daphne_brain/requirements.txt
channels~=3.0.3
channels-redis~=3.2.0
dateparser~=1.0.0
Django~=3.2.2
django-cors-headers~=3.7.0
django-filter~=2.4.0
djangorestframework~=3.12.2
neo4j==4.2.1
neo4j-driver==4.2.1
numpy~=1.19.2
pandas~=1.2.3
pika~=1.2.0
psycopg2-binary~=2.8.6
python-Levenshtein~=0.12.2
redis~=3.5.3
schedule==1.1.0
scipy~=1.6.1
# spacy~=3.0.3
# statsmodels~=0.12.2
SQLAlchemy==1.4.14
# tensorflow~=2.4.1
thrift~=0.13.0
xlrd~=2.0.1
asgiref~=3.3.4
scikit-learn~=0.24.2
pslpython~=2.2.2
PyYAML~=5.4.1
openai
langchain
langchain-community
python-dotenv[cli]
typing_extensions
langchain_neo4j